themed the popup that comes up when you hover on the activity entries
this one was a bit difficult because I had to switch to the Sources tab, press F8 to pause the JS, then pick the element to theme, and then I was able to copy the element class.
Anything that only appears when another element is hovered on will do that.